fix: keep httpapi instance reloads in layer store
