+++ packages/opencode/src/project/instance-context.ts
@@ -1,0 +1,10 @@
+import { LocalContext } from "@/util/local-context"
+import type * as Project from "./project"
+
+export interface InstanceContext {
+  directory: string
+  worktree: string
+  project: Project.Info
+}
+
+export const context = LocalContext.create<InstanceContext>("instance")
