+++ packages/opencode/src/server/routes/instance/httpapi/lifecycle.ts
@@ -2,11 +2,13 @@ import type { WorkspaceID } from "@/control-plane/schema"
 import { WorkspaceContext } from "@/control-plane/workspace-context"
 import { WorkspaceRef } from "@/effect/instance-ref"
 import { Instance, type InstanceContext } from "@/project/instance"
+import { InstanceStore } from "@/project/instance-store"
 import { Effect } from "effect"
 import { HttpEffect, HttpMiddleware, HttpServerRequest } from "effect/unstable/http"
 
 type MarkedInstance = {
   ctx: InstanceContext
+  store: InstanceStore.Interface
   workspaceID?: WorkspaceID
 }
 
@@ -17,17 +19,17 @@ const disposeAfterResponse = new WeakMap<object, MarkedInstance>()
 
 const mark = (ctx: InstanceContext) =>
   Effect.gen(function* () {
-    return { ctx, workspaceID: yield* WorkspaceRef }
+    return { ctx, store: yield* InstanceStore.Service, workspaceID: yield* WorkspaceRef }
   })
 
-// Instance.dispose/reload still publish events through legacy ALS helpers.
+// InstanceStore lifecycle operations still publish events through legacy ALS helpers.
 // Effect request handlers carry these values in services, so bridge them back
 // into the legacy contexts only around the lifecycle operation.
-const restoreMarked = <A>(marked: MarkedInstance, fn: () => A) =>
+const restoreMarked = <A>(marked: MarkedInstance, effect: Effect.Effect<A>) =>
   Effect.promise(() =>
     WorkspaceContext.provide({
       workspaceID: marked.workspaceID,
-      fn: () => Instance.restore(marked.ctx, fn),
+      fn: () => Instance.restore(marked.ctx, () => Effect.runPromise(effect)),
     }),
   )
 
@@ -43,11 +45,11 @@ export const markInstanceForDisposal = (ctx: InstanceContext) =>
     )
   })
 
-export const markInstanceForReload = (ctx: InstanceContext, next: Parameters<typeof Instance.reload>[0]) =>
+export const markInstanceForReload = (ctx: InstanceContext, next: InstanceStore.LoadInput) =>
   Effect.gen(function* () {
     const marked = yield* mark(ctx)
     return yield* HttpEffect.appendPreResponseHandler((_request, response) =>
-      Effect.as(Effect.uninterruptible(restoreMarked(marked, () => Instance.reload(next))), response),
+      Effect.as(Effect.uninterruptible(restoreMarked(marked, marked.store.reload(next))), response),
     )
   })
 
@@ -58,6 +60,6 @@ export const disposeMiddleware: HttpMiddleware.HttpMiddleware = (effect) =>
     const marked = disposeAfterResponse.get(request.source)
     if (!marked) return response
     disposeAfterResponse.delete(request.source)
-    yield* Effect.uninterruptible(restoreMarked(marked, () => Instance.dispose()))
+    yield* Effect.uninterruptible(restoreMarked(marked, marked.store.dispose(marked.ctx)))
     return response
   })
